requirements.txt
altgraph==0.17.3
anyio==3.6.2
auto-py-to-exe==2.33.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
click==8.1.3
colorama==0.4.6
Eel==0.16.0
fastapi==0.94.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pefile==2023.2.7
psutil==5.9.4
pycparser==2.21
pydantic==1.10.6
pygame-ce==2.3.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pyparsing==3.0.9
pywin32-ctypes==0.2.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
websocket-client==1.5.1
websockets==10.4
whichcraft==0.6.1
zope.event==4.6
zope.interface==6.0
